Produce a plot for a feature_calculations object

Produce a plot for a feature_calculations object

Usage

# S3 method for feature_calculations
plot(
  x,
  type = c("quality", "matrix", "cor"),
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax"),
  clust_method = c("average", "ward.D", "ward.D2", "single", "complete", "mcquitty",
    "median", "centroid"),
  cor_method = c("pearson", "spearman"),
  ...
)

Arguments

x

the feature_calculations object containing the raw feature matrix produced by calculate_features

type

character specifying the type of plot to draw. Defaults to "quality"

method

character specifying a rescaling/normalising method to apply if type = "matrix" or if type = "cor". Defaults to "z-score"

clust_method

character specifying the hierarchical clustering method to use if type = "matrix" or if type = "cor". Defaults to "average"

cor_method

character specifying the correlation method to use if type = "cor". Defaults to "pearson"

...

Arguments to be passed to methods

Value

object of class ggplot that contains the graphic

Author

Trent Henderson