Project a feature matrix into a low dimensional representation using PCA or t-SNE

Project a feature matrix into a low dimensional representation using PCA or t-SNE

Usage

reduce_dims(
  data,
  method = c("z-score", "Sigmoid", "RobustSigmoid", "MinMax"),
  low_dim_method = c("PCA", "t-SNE"),
  perplexity = 30,
  seed = 123
)

Arguments

data

the feature_calculations object containing the raw feature matrix produced by calculate_features

method
<<<<<<< HEAD

character specifying a rescaling/normalising method to apply. Defaults to "z-score"

low_dim_method

character specifying the low dimensional embedding method to use. Defaults to "PCA"

perplexity

integer denoting the perplexity hyperparameter to use if t-SNE algorithm is selected. Defaults to 30

seed

integer to fix R's random number generator to ensure reproducibility. Defaults to 123

=======

a rescaling/normalising method to apply. Defaults to "z-score"

low_dim_method

the low dimensional embedding method to use. Defaults to "PCA"

perplexity

the perplexity hyperparameter to use if t-SNE algorithm is selected. Defaults to 30

seed

fixed number for R's random number generator to ensure reproducibility

>>>>>>> main

Value

object of class low_dimension

Author

Trent Henderson